보드 등록 및 수정 , 불러오기 , 삭제 구현

+++ teamKMS/src/main/java/com/devworker/kms/service/site/SiteService.java
@@ -10,17 +10,22 @@ import com.devworker.kms.dto.site.SiteDto;
 import com.devworker.kms.entity.site.ProjectBoardDao;
 import com.devworker.kms.entity.site.ProjectDao;
 import com.devworker.kms.entity.site.SiteDao;
+import com.devworker.kms.exception.NotAllowException;
+import com.devworker.kms.exception.NotExistException;
 import com.devworker.kms.repo.site.ProjectBoardRepo;
 import com.devworker.kms.repo.site.ProjectRepo;
 import com.devworker.kms.repo.site.SiteRepo;
 import com.devworker.kms.util.AclUtil;
+import com.devworker.kms.util.CommonUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -53,8 +58,7 @@ public class SiteService {
     public Page<BoardDto> getProjectBoards(int menuId, int siteId, int projectId, Pageable pageable) {
          Page<ProjectBoardDto> projectBoardDtoLIst=projectBoardRepo.getSiteProjects(projectId, pageable).map(ProjectBoardDao::getDto);
          List<Long> collect =new ArrayList<>();
-
-        for(ProjectBoardDto  projectDto : projectBoardDtoLIst)
+          for(ProjectBoardDto  projectDto : projectBoardDtoLIst)
             collect.add(projectDto.getBoardId());
         return boardWriteService.getPages(collect, pageable);
     }
@@ -67,22 +71,34 @@ public class SiteService {
         projectRepo.deleteById(projectId);
     }
 
-    public void addBoard(BoardDetailDto dto, int siteId, int projectId) {
+    public void addBoard(ProjectBoardDto projectBoardDto) {
         //TODO::
-    	ProjectBoardDao projectBoardDao=new ProjectBoardDao();
-    	projectBoardDao.setBoardId((boardWriteService.register(dto)));
-    	projectBoardDao.setProjectId(projectId);
-    	
-       projectBoardRepo.save(projectBoardDao);
+        projectBoardDto.getBoardDetailDto().setRegDate(LocalDateTime.now());
+        projectBoardDto.getBoardDetailDto().setUpdDate(LocalDateTime.now());
+        projectBoardDto.getBoardDetailDto().setUserId(CommonUtil.getCurrentUser());
+        projectBoardDto.setBoardId((boardWriteService.register(projectBoardDto.getBoardDetailDto())));
+
+       projectBoardRepo.save(projectBoardDto.getDao());
+    }
+    public void editBoard(ProjectBoardDto projectBoardDto) {
+        //TODO::
+        projectBoardDto.getBoardDetailDto().setUpdDate(LocalDateTime.now());
+        projectBoardDto.getBoardDetailDto().setUserId(CommonUtil.getCurrentUser());
+
+        boardWriteService.register(projectBoardDto.getBoardDetailDto());
     }
 
-    public void deleteBoard(int siteId, int projectId, int boardId) {
+    public void deleteBoard(int siteId, int projectId, long boardId) {
         //TODO::
-    	projectBoardRepo.deleteById(boardId);
+        BoardDetailDto boardDetailDto=boardWriteService.getBoard(boardId);
+        if(!AclUtil.checkPermission(boardDetailDto.getUserId() , PermissionType.DELETESOL))
+            throw new NotAllowException("You Have not Delete Project Board Permission");
+       	projectBoardRepo.deleteById(boardId);
+        boardWriteService.delete(boardId);
     }
     public BoardDetailDto getSiteProjectBoardById(Long id) {
-       /* Optional<SiteDao> siteDao =  projectRepo.findById(id);
-        siteDao.orElseThrow(() -> new NotExistException("Site Board Not Found"));*/
+        Optional<ProjectBoardDao> projectBoardDao =  projectBoardRepo.findByBoardId(id);
+        projectBoardDao.orElseThrow(() -> new NotExistException("Site Board Not Found"));
         BoardDetailDto dto = boardWriteService.getBoard(id);
         dto.setReadOnly(!AclUtil.checkPermission(dto.getUserId() , PermissionType.MODIFYSOL));
         dto.setHits(dto.getHits() +1);
